fix(planner): first planner error aborts the process
Stream poisoning is gone per review: a planner that produced an invalid
timeline has nothing safe left to do, so the first dispatch or shape error
logs a fatal line and calls process::abort() — klippy dies with it and the
MCUs shut down on host disconnect. Deletes CommitOutcome, the error slot,
check_error, and every stream_poisoned guard; drops the poisoning tests.

REPLAN_WARN_BUDGET_US is computed from LEAD - SAFETY_MARGIN instead of
hardcoded. anchor.rs unit tests move to anchor/tests.rs per convention.
Comment sweep across the PR's files per CLAUDE.md comment policy.
